refactor event_builder
